ARC-0003: Support for Integrity Properties
Currently, there is no easy way to check the integrity of files
pointed by the metadata JSON. In particular, when buying an
NFT, the seller may then change for example the image of the
NFT without the buyer having anyway to prove that the image
actually changed.

This PR proposes to fix this issue by adding additional integrity
properties.