ARC-0003: Adding MIME types and OpenSea fields
This PR is meant to try to provide clearer guidelines
so that clients such as marketplaces can more easily
display NFTs, following some discussions in
https://github.com/algorandfoundation/ARCs/issues/3

Concretely:
1. It adds the fields `external_url`, `animation_url` and
`background_color` used on OpenSea.
2. It clarifies the use of those fields in the example of
a song.
3. It adds `_mimetype` fields that are optional
(for compatiblity with OpenSea) but highly recommended
4. It switched to the term `fields` for JSON fields
to avoid confusion with properties/attributes of the
asset.

Rationale:
1. Keep the fields in the "root" of the JSON metadata
file (as opposed to the properties) limited to
what is required for preview in a markeplace/gallery.
In particular `image` and `animation_url` can be,
for some assets, just preview, as opposed to the
full asset.
Remark for example that the field `image` restricts
a lot the maximum dimension of the image, so that
high-quality photography could not fit this field.
2. Full media files (e.g., full songs, high-quality photo),
licenses, and other such properties
are moved to properties and are left out of scope of this
proposal. Future proposals may try to standardize these
properties, although a global cross-chain standard
would be even more beneficial.

Points of dicussion:
1. Should MIME types be mandatory? This would break
backward compatibility with ERC-1155 and make most
existing NFT metadata invalid in the context of ARC-3.
The proposed solution is that clients may decide to
guess the MIME type from the extension of the file
(or from a HEAD query) when not available. The former
is what OpenSea seems to be doing.
This is much more brittle, hence the strong recommendation
for specifying MIME types.
2. Should URI properties (i.e., inside the `properties` field
and not at the root), should also follow the same rules.
In particular, should we allow `{id}` there? It may actually
be an issue as marketplace may not be able to recognize
which fields are URI.
3. Is this proposal too restrictive and should we finalize
ARC-0003 without it completely, sticking to being close
to ERC-1155 (with integrity fields and extra metadata though)
(In case this proposal is rejected, we may want to keep
the renaming of properties into fields.)