Update arc-0048.md
Improving the Measure (ARC48) to propose for the community to allow more flexibility on defi priorities. The projects would then be able to reward users and work on user acquisition.
It is an improvement to reduce admin burden and economically insignificant (smaller) reward amounts. And, all rewards utilisation will be disclosed to the community in whole (publicly and timely) and, wholly devoted to protocol provision, user rewards, user engagement.